Separate the RequestTracer static methods from the RequestHandler instance.
